+++ ios/Classes/Constants.swift
@@ -9,4 +9,8 @@ struct Constants {
     
     static let EXTRA_MEASUREMENT_UUID: String = "EXTRA_MEASUREMENT_UUID"
     static let EXTRA_LICENSE_KEY: String = "EXTRA_LICENSE_KEY"
+    
+    static let ERROR_CODE_PLUGIN_NOT_ATTACHED_TO_ACTIVITY: String = "1000"
+    static let ERROR_CODE_SDK_INITIALIZATION_FAILED: String = "1001"
+    static let ERROR_CODE_GENERIC_EXCEPTION: String = "1003"
 }