+++ test/anyline_tire_tread_plugin_test.dart
@@ -1,5 +1,5 @@
 import 'package:anyline_tire_tread_plugin/src/models/scan_options.dart';
-import 'package:anyline_tire_tread_plugin/src/scanning_event.dart';
+import 'package:anyline_tire_tread_plugin/src/scan_event.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:anyline_tire_tread_plugin/src/anyline_tire_tread_plugin_platform_interface.dart';
 import 'package:anyline_tire_tread_plugin/src/anyline_tire_tread_plugin_method_channel.dart';